Move seed runner to its own namespace under lib
